requirements.txt
click==8.1.3
Flask==2.1.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
uWSGI==2.0.20
Werkzeug==2.2.1
zipp==3.8.1
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
Flask==2.1.3
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.3
Pillow==9.2.0
requests==2.28.1
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
typing_extensions==4.3.0
urllib3==1.26.12
uWSGI==2.0.20
Werkzeug==2.2.1
zipp==3.8.1
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
cycler==0.11.0
Flask==2.1.3
fonttools==4.37.1
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
numpy==1.23.3
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.1
six==1.16.0
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
typing_extensions==4.3.0
urllib3==1.26.12
uWSGI==2.0.20
Werkzeug==2.2.1
zipp==3.8.1
